requirements.txt
# Resume PDF Generator Requirements
# Install with: pip install -r requirements.txt

# Required for PDF generation
reportlab>=4.0.0

# Optional for automatic duration calculation
python-dateutil>=2.8.0
